proof of concept of ad hoc workflow
The idea behind this is some way to launch a number of jobs with 1
request, the client could then subscribe to websocket channel and get
updates on status of the nodes or at least just poll this workflow until
individual job nodes are done and then reach into the jobs. This is all
to provide some interface that allows clients to get information they
need without barraging controller with many many requests.

First pass does not attempt to implement any workflow structure other
than flat, maybe not even necessary if only goal is a bulk job launch
endpoint. Even with the bulk job launch endpoint maybe would want
"identifier"s on the workflow nodes to be able to tell the jobs apart in
websocket messages
or the api.

This first approach probably doesn't check for RBAC permissions on the
workflow nodes correctly. Maybe you would want a whole new role for
ability to launch adhoc workflows.

Also, have not optimized database access. Observed 29 queries with 1
workflow node with both ujt and inventory, and ~1500 queries with 100
nodes with ujt and inventory.

At some point if we can't optimize it we'd have to have some kind of
model like AdHocWorkflow with the proto-nodes in it and defer creation to
some background task, but then we wouldn't get the RBAC check at launch
time -- which I think is desirable. I'd prefer to optimize this and
limit size of AdHocWorkflows.

example payload:
```
{
"name": "foobar example bulk launch",
"workflow_job_nodes": [
{"unified_job_template": 7, "inventory": 1},
{"unified_job_template": 7}
]
}
```

optimize queries, down to 30ish queries from 1500

to make 93 nodes with ujt and inventory defined, down to 30 queries from
up around 1500

did discover that how I was doing credentials doesnt work -- right now
is no-op to pass list of credentials, need to figure out how to deal
with the fact that you can't pass list of credentials to WorkflowJobNode
init and you can't call wf_node.credentials.set() until after the node
has an id (e.g., we've called create).

Maybe if we had an identifier we could go back and add the credentials
after...

make credentials work

need to set these after node is created afaict

check roles for workflow node related objects

Found this method in access.py, not 100% sure these are the roles we
want or if this is best way, but seems to be working

no model behind this view to check permissions on

Tested with regular user, and need this to at least see the endpoint and
POST to it -- but got expected behavior as far as not launching the
workflow when user lacked permissions on the objects in the workflow
nodes.

update comments

now project updates work too

inventory source updates not working yet

use a set comprehension at jrb's prompting

change interface to give clue on what job type

new interface allows us to know what kind of job we are launching,
because we need this for the RBAC queries to check access. If we know
what job types we are looking for, we can skip some queries.

Example POST payload:

{
    "name": "foobar example bulk launch",
    "workflow_job_nodes": [
        {"workflow_job_template_id": 224},
        {"workflow_job_template_id": 225},
        {"workflow_job_template_id": 226},
        {"workflow_job_template_id": 227},
        {"job_template_id": 7},
        {"job_template_id": 7, "inventory": 1, "credentials": [18]},
        {"project_id": 6}
    ]
}

fixup comments and unused import

show url in root v2 urls

get inventory source updates working

also be consistent with unpacking of values_list with set comprehension

skip rbac checks if is superuser

also fixup things for normal user seeing view